Split snapshot integration tests into thematic submodules
Split the monolithic snapshot.rs (~3100 lines, 75 tests) into a
snapshot/ directory with 10 thematic files matching the pattern used
by other extension test modules (tags/, fixtures/).

Also inlines verbose command construction patterns where a `let mut cmd`
+ separate `.arg()` call can be expressed as a single chained expression.
