+++ crates/karva/tests/it/extensions/snapshot/mod.rs
@@ -1,0 +1,10 @@
+mod basic;
+mod commands;
+mod content;
+mod delete;
+mod filters;
+mod inline;
+mod json;
+mod named;
+mod prune;
+mod review;
